Exp_struct: Rearrange for adding dict category number
... while still using 32 bytes (on most 64-bit CPUs).

+++ link-grammar/dict-common/print-dict.c
@@ -403,7 +403,7 @@ static const char *stringify_Exp_tag(Exp *e, Dictionary dict)
 				if (dict == NULL)
 				{
 					snprintf(tag_info, sizeof(tag_info), " dialect_tag=%u",
-					         e->tag_id);
+					         (unsigned int)e->tag_id);
 				}
 				else
 				{
